backend/api/serializers.py
from rest_framework import serializers

class UserPublicSerializer(serializers.Serializer):
    id = serializers.IntegerField(read_only = True)
    username = serializers.CharField(read_only = True)

class HistoricalRecordSerializer(serializers.ListField):
    child = serializers.DictField()

    def to_representation(self, data):
        return super().to_representation(data.values())